requirements.txt
opencv-python
numpy
Pillow
mediapipe
ultralytics